makes changes to client-side

+++ client-side/sync.py
@@ -37,6 +37,15 @@ def delete_post(ur, dat,dat1,b,s):
         s.post(url=ur, data=dat)
     elif b==2:
         s.delete(url=ur,data=dat1)
+    elif b==3:
+        if os.path.exists(ur):
+            os.remove(ur)
+    elif b==4:
+        if os.path.exists(ur):
+            os.remove(ur)
+        print()
+    elif b==5:
+        print()
 
 def md5(fname):
     hash_md5 = hashlib.md5()
@@ -126,6 +135,7 @@ def status(user, pas, userid, rootDir, enc_type, server_url):
 
 
 def sync(user, pas, userid, rootDir, enc_type, server_url):
+    client_server=input('Change files on client or server? (c or s): ')
     p = Path(rootDir)
     dirname = str(p.name)
     s = requests.Session()
@@ -153,20 +163,26 @@ def sync(user, pas, userid, rootDir, enc_type, server_url):
             if (r2.ok):
                 dicti = r2.json()
                 if dicti['md5code'] != md5(complete_path):
-                    encrypt.encrypt(complete_path, enc_type,pas)
-                    with open(complete_path + '.' + enc_type + 'en', 'rb') as con:
-                        content = con.read()
-                    if os.path.exists(complete_path + '.' + enc_type + 'en'):
-                        os.remove(complete_path + '.' + enc_type + 'en')
-                    dicti['fileContent']=content
-                    dicti['md5code']= md5(complete_path)
-                    dicti['username']=user
-                    del dicti['modifiedTime']
-                    del dicti['pk']
-                    print('Replacing ' + dirname + '/' + relFile + '/')
-                    dic={'owner': int(userid),'username':user,'password':pas}
-                    progress_bar(str(server_url+'/user/' + user + '/data/' + dirname +'.'+enc_type+'en'+ '/' + relFile) + '/',
-                                 dicti,dic, 1, s)
+                    if client_server == 's':
+                        encrypt.encrypt(complete_path, enc_type, pas)
+                        with open(complete_path + '.' + enc_type + 'en', 'rb') as con:
+                            content = con.read()
+                        if os.path.exists(complete_path + '.' + enc_type + 'en'):
+                            os.remove(complete_path + '.' + enc_type + 'en')
+                        dicti['fileContent'] = content
+                        dicti['md5code'] = md5(complete_path)
+                        dicti['username'] = user
+                        del dicti['modifiedTime']
+                        del dicti['pk']
+                        print('Replacing on server ' + dirname + '/' + relFile + '/')
+                        dic = {'owner': int(userid), 'username': user, 'password': pas}
+                        progress_bar(str(
+                            server_url + '/user/' + user + '/data/' + dirname + '.' + enc_type + 'en' + '/' + relFile) + '/',
+                                     dicti, dic, 1, s)
+                    else:
+                        print('Replacing on client ' + dirname + '/' + relFile + '/')
+                        progress_bar(complete_path,{},{},4,s)
+
             else:
                 dirlist = os.path.normpath(dirname + '/' + relFile)
                 dirlist = dirlist.split(os.sep)
@@ -178,69 +194,85 @@ def sync(user, pas, userid, rootDir, enc_type, server_url):
                         pat1 = pat + str(dirlist[i]) + '.' + enc_type + 'en' + '/'
                         r2 = s.get(server_url + '/user/' + user + '/contents/' + pat1, data={'owner': int(userid)})
                         if not r2.ok:
-                            r3= s.get(server_url + '/user/' + user + '/contents/' + pat, data={'owner': int(userid)})
-                            data2 = r3.json()
-                            parentid = data2['pk']
-                            owner = int(userid)
-                            username = user
-                            name=str(dirlist[i])+'.'+enc_type+'en'
-                            md5code = md5(complete_path)
-                            dorf = 'f'
-                            pathLineage = pat + str(name) + '.' + enc_type + 'en' + '/'
-                            encrypt.encrypt(complete_path, enc_type, pas)
-                            with open(complete_path + '.' + enc_type + 'en', 'rb') as con:
-                                content = con.read()
-                            if os.path.exists(complete_path + '.' + enc_type + 'en'):
-                                os.remove(complete_path + '.' + enc_type + 'en')
-                            fileContent = content
-                            dicti = {'owner': owner, 'parentId': int(parentid), 'name': name,
-                                     'pathLineage': pathLineage,
-                                     'dorf': dorf,
-                                     'fileContent': fileContent, 'md5code': md5code, 'username': username}
-                            dic = {}
-                            print('Adding ' + pathLineage)
-                            progress_bar(server_url + '/user/' + user + '/data/' + pathLineage,
-                                         dicti, dic, 0, s)
+                            if client_server == 's':
+                                r3 = s.get(server_url + '/user/' + user + '/contents/' + pat,
+                                           data={'owner': int(userid)})
+                                data2 = r3.json()
+                                parentid = data2['pk']
+                                owner = int(userid)
+                                username = user
+                                name = str(dirlist[i]) + '.' + enc_type + 'en'
+                                md5code = md5(complete_path)
+                                dorf = 'f'
+                                pathLineage = pat1
+                                encrypt.encrypt(complete_path, enc_type, pas)
+                                with open(complete_path + '.' + enc_type + 'en', 'rb') as con:
+                                    content = con.read()
+                                if os.path.exists(complete_path + '.' + enc_type + 'en'):
+                                    os.remove(complete_path + '.' + enc_type + 'en')
+                                fileContent = content
+                                dicti = {'owner': owner, 'parentId': int(parentid), 'name': name,
+                                         'pathLineage': pathLineage,
+                                         'dorf': dorf,
+                                         'fileContent': fileContent, 'md5code': md5code, 'username': username}
+                                dic = {}
+                                print('Adding to server ' + pathLineage)
+                                progress_bar(server_url + '/user/' + user + '/data/' + pathLineage,
+                                             dicti, dic, 0, s)
+                            else:
+                                print('Deleting from client ' + pat + str(dirlist[i]) + '/' )
+                                progress_bar(complete_path,{},{},3,s)
                     else:
+                        pat2=''
                         for j in range(i):
+                            pat2=pat2+str(dirlist[j+1]) + '/'
                             pat = pat + str(dirlist[j]) + '/'
                         pat1 = pat + str(dirlist[i]) + '/'
                         r2 = s.get(server_url + '/user/' + user + '/contents/' + pat1, data={'owner': int(userid)})
                         if not r2.ok:
-                            r3= s.get(server_url + '/user/' + user + '/contents/' + pat, data={'owner': int(userid)})
-                            if r3.ok:
-                                data2 = r3.json()
-                                parentid = data2['pk']
-                            else:
-                                parentid=0
+                            if client_server == 's':
+                                r3 = s.get(server_url + '/user/' + user + '/contents/' + pat,
+                                           data={'owner': int(userid)})
+                                if r3.ok:
+                                    data2 = r3.json()
+                                    parentid = data2['pk']
+                                else:
+                                    parentid = 0
 
-                            owner = int(userid)
-                            username = user
-                            name=str(dirlist[i])
-                            md5code ='-'
-                            dorf = 'd'
-                            pathLineage = pat + str(name)+'/'
-                            fileContent = '-'
-                            dicti = {'owner': owner, 'parentId': int(parentid), 'name': name,
-                                     'pathLineage': pathLineage,
-                                     'dorf': dorf,
-                                     'fileContent': fileContent, 'md5code': md5code, 'username': username}
-                            dic = {}
-                            print('Adding ' + pathLineage)
-                            progress_bar(server_url + '/user/' + user + '/data/' + pathLineage,
-                                         dicti, dic, 0, s)
+                                owner = int(userid)
+                                username = user
+                                name = str(dirlist[i])
+                                md5code = '-'
+                                dorf = 'd'
+                                pathLineage = pat + str(name) + '/'
+                                fileContent = '-'
+                                dicti = {'owner': owner, 'parentId': int(parentid), 'name': name,
+                                         'pathLineage': pathLineage,
+                                         'dorf': dorf,
+                                         'fileContent': fileContent, 'md5code': md5code, 'username': username}
+                                dic = {}
+                                print('Adding to server ' + pathLineage)
+                                progress_bar(server_url + '/user/' + user + '/data/' + pathLineage,
+                                             dicti, dic, 0, s)
+                            else:
+                                print('Deleting from client ' + pat + str(dirlist[i]) + '/')
+                                progress_bar(os.path.join(rootDir, pat2), {}, {}, 3, s)
+                                break
 
     r = s.get(url=server_url + '/user/' + user + '/allfiles/'+str(p.name), data={'owner': int(userid)})
     dat=r.json()
     for pat in dat:
         fil=pat['pathLineage']
         if not fil in file_dir_list:
-            dic = {'owner': int(userid), 'username': user, 'password': pas}
-            dicti = {}
-            l=len(fil)
-            if fil[l-7:l-1]==str('.'+enc_type+'en'):
-                print('Deleting ' + fil[0:-7]+'/' )
-                progress_bar(server_url + '/user/' + user + '/data/' +fil[0:-7]+'/', dicti, dic, 2, s)
+            if client_server == 's':
+                dic = {'owner': int(userid), 'username': user, 'password': pas}
+                dicti = {}
+                l = len(fil)
+                if fil[l - 7:l - 1] == str('.' + enc_type + 'en'):
+                    print('Deleting from server ' + fil[0:-7] + '/')
+                    progress_bar(server_url + '/user/' + user + '/data/' + fil[0:-7] + '/', dicti, dic, 2, s)
+                else:
+                    print('Deleting from server ' + fil)
+                    progress_bar(server_url + '/user/' + user + '/data/' + fil, dicti, dic, 2, s)
             else:
-                print('Deleting ' + fil)
-                progress_bar(server_url + '/user/' + user + '/data/' + fil, dicti, dic, 2, s)
+                print()
